requirements.txt
Django==3.2.0
djangorestframework==3.12.0
mysqlclient==2.0.3

djangorestframework-simplejwt==5.3.0

django-filter==23.3

mysqlclient==2.0.3

